+++ lib/src/helpers/events/events.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../../web.dart';
+import '../../../web.dart';
 import 'providers.dart';
 import 'streams.dart';
